refactor: filter_search_sheet 폴더 추가로 위치 변경
