+++ lib/presentation/component/component_test_screen.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:recipe_app/core/ui/color_styles.dart';
 import 'package:recipe_app/core/ui/text_styles.dart';
-import 'package:recipe_app/presentation/component/filter_button.dart';
+import 'package:recipe_app/presentation/component/filter_search_sheet/filter_button.dart';
 import 'package:recipe_app/presentation/component/input_field_widget.dart';
-import 'package:recipe_app/presentation/component/rating_button.dart';
+import 'package:recipe_app/presentation/component/filter_search_sheet/rating_button.dart';
 import 'package:recipe_app/presentation/component/tabs_widget.dart';
 
 import 'button_widget.dart';
